requirements.txt
annotated-types==0.7.0
anyio==4.12.1
babel==2.17.0
bcrypt==5.0.0
beanie==2.0.0
black==23.1.0
blinker==1.9.0
cachelib==0.13.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
fastapi-pagination==0.15.3
Flask==3.1.2
flask-babel==4.0.0
Flask-Caching==2.3.1
flask-mongoengine-3==1.1.0
h11==0.16.0
httptools==0.7.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lazy-model==0.3.0
livereload==2.7.1
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mongoengine==0.29.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
pendulum==3.1.0
platformdirs==4.5.1
pyasn1==0.6.1
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
starlette==0.47.3
tornado==6.5.4
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchdog==2.2.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.4
